seat: Implement a dummy touch grab
On devices other than the Librem 5, such as the PinePhone, the touch
panel is still enabled when the screen is blanked. This results in touch
events being sent to the shell even with a blank screen, resulting in
unsollicited application launch and other unwanted behavior.

In order to avoid that, we need a way to disable touch events when
required. The best way to achieve this is to setup a dummy touch grab,
which will capture all touch events, but not do anything with them,
which is the point of the new `roots_seat_toggle_touch` function this
commit introduces.